\name{association.corrplot}
\alias{association.corrplot}
\title{association.corrplot}
\description{represent association between components and annotation in a corrplot }
\usage{association.corrplot(pvaltab_uni, pvaltab_multi)}
\arguments{
\item{pvaltab_uni}{pvalue of univariate analysis}
\item{pvaltab_multi}{pvalue of multivariate analysis}
}
\value{NA
}
\author{Lea Meunier}
\keyword{representation}
