\name{enrich.test}
\alias{enrich.test}
\title{enrich.test}
\description{compute the enrichment of CpG feature }
\usage{enrich.test(CpG_select, column, CpG_feature) }
\arguments{
\item{CpG_select}{CpG selection }
\item{column}{feature to compute enrichment}
\item{gene_table}{table with CpG feature}
}
\value{table with enrichment results
}
\author{Lea Meunier}
\keyword{test}
