\name{enrich.CpG.feature}
\alias{enrich.CpG.feature}
\title{enrich.CpG.feature}
\description{comptute enrichement of CpG feature and generate repressentation}
\usage{enrich.CpG.feature(MC_object, MC_contrib_CpG, output.directory, CpG_feature)}
\arguments{
\item{MC_object}{Object return by mc.extract}
\item{MC_contrib_CpG}{object return by mc.active.CpG with most contributing CpGs for each component}
\item{output.directory}{path to save output}
}
\author{Lea Meunier}
\keyword{visualisation}
