\name{mc.change}
\alias{mc.change}
\title{mc.change}
\description{required}
\usage{mc.change(MC_object, MC_active_sample, MC_contrib_CpG, bval, ref, output.directory)}
\arguments{
\item{MC_object}{Object return by mc.extract}
\item{MC_active_sample}{object return by mc.activ.sample with active sample in each component}
\item{MC_contrib_CpG}{object return by mc.active.CpG with most contributing CpGs for each component}
\item{bval}{bval methylation data}
\item{ref}{liste of samples use like reference}
\item{output.directory}{path to save output}
}
\author{Lea Meunier}
\keyword{visualisation}
